torchseq/utils/seed.py
import os
import random

import numpy as np
import torch


def set_seed(seed=1029):
    random.seed(seed)
    os.environ["PYTHONHASHSEED"] = str(seed)
    os.environ["__TORCHAQSEED"] = str(seed)
    np.random.seed(seed)
    torch.manual_seed(seed)
    torch.cuda.manual_seed(seed)
    torch.cuda.manual_seed_all(seed)  # if you are using multi-GPU.
    torch.backends.cudnn.benchmark = False
    torch.backends.cudnn.deterministic = True


def init_worker(worker_id):
    set_seed(int(os.environ["__TORCHAQSEED"]))
