requirements.txt
absl-py==0.9.0
tensorboard==2.2.2
torch==1.5.0
tqdm
# Tokenizermb
nltk==3.5.0
transformers==2.9.0
tokenizers==0.7.0
jsonlines
sacrebleu

pytest
pytest-cov
codecov
flake8
black